Split userMenu component into its own file
