+++ lib/components/Header/Header.styles.js
@@ -107,7 +107,7 @@ export const UserMenuContainer = styled(Box)((props) =>
   })
 );
 
-export const UserMenu = styled(Box)((props) =>
+export const UserMenuContent = styled(Box)((props) =>
   css({
     position: "absolute",
     gap: "s",
