+++ lib/components/Header/index.js
@@ -1,20 +1,9 @@
-import React, { useState, useRef, useEffect } from "react";
+import React from "react";
 import PropTypes from "prop-types";
-import Avatar from "../Avatar";
-import Icon from "../Icon";
-import Divider from "../Divider";
 import Flex, { FlexItem } from "../Flex";
 import { Small } from "../Typography";
-import {
-  Bar,
-  AppName,
-  SearchContainer,
-  UserMenuToggle,
-  AvatarBorder,
-  UserMenuContainer,
-  UserMenu
-} from "./Header.styles";
-
+import { Bar, AppName, SearchContainer } from "./Header.styles";
+import UserMenu from "./UserMenu";
 /**
  * Header component for top of app.
  **/
@@ -26,31 +15,13 @@ const Header = ({
   avatarInitials,
   searchComponent,
   dataTestId,
-  variant,
   currentWorkspace,
   logo,
   userMenuContent
 }) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const menuRef = useRef(null);
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (menuRef.current && !menuRef.current.contains(event.target)) {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
   return (
-    <Bar dataTestId={dataTestId} variant={variant}>
-      <AppName variant={variant}>
+    <Bar dataTestId={dataTestId}>
+      <AppName>
         {logo && (
           <Flex alignItems="center" justifyContent="center">
             {logo}
@@ -71,57 +42,13 @@ const Header = ({
       </AppName>
       {searchComponent && <SearchContainer>{searchComponent}</SearchContainer>}
       {userName && (
-        <UserMenuContainer
-          ref={menuRef}
-          p="2px"
-          pr="s"
-          bg="white"
-          width="auto"
-          height="38px"
-          border="solid 1px"
-          borderColor="greyLight"
-          shadow="default"
-          display="flex"
-        >
-          <UserMenuToggle onClick={toggleMenu}>
-            <AvatarBorder borderRadius="50%" p="2px">
-              <Avatar
-                customSize="28px"
-                image={avatarSource}
-                imageAlt={avatarAlt}
-                initials={avatarInitials}
-              />
-            </AvatarBorder>
-            <Icon
-              icon={["fas", "chevron-down"]}
-              rotation={isOpen ? 180 : 0}
-              color="greyDark"
-              ml="s"
-            />
-          </UserMenuToggle>
-          <UserMenu
-            p="s"
-            bg="white"
-            width="fit-content"
-            maxWidth="260px"
-            maxHeight="430px"
-            border="solid 1px"
-            borderRadius="2"
-            display="flex"
-            flexDirection="column"
-            alignItems="flex-start"
-            justifyContent="flex-start"
-            borderColor="greyLight"
-            shadow="default"
-            isOpen={isOpen}
-          >
-            <Small color="greyDarker" fontSize="0">
-              {userName}
-            </Small>
-            <Divider light />
-            {userMenuContent}
-          </UserMenu>
-        </UserMenuContainer>
+        <UserMenu
+          userName={userName}
+          avatarSource={avatarSource}
+          avatarAlt={avatarAlt}
+          avatarInitials={avatarInitials}
+          userMenuContent={userMenuContent}
+        />
       )}
     </Bar>
   );
@@ -134,7 +61,7 @@ Header.propTypes = {
   appName: PropTypes.node,
   /** Current active workspace name. Can be a plain text string or a component */
   currentWorkspace: PropTypes.node,
-  /** Text for username. */
+  /** Text for logged in user name and email. */
   userName: PropTypes.node,
   /** Source path for avatar image. */
   avatarSource: PropTypes.node,
@@ -146,8 +73,6 @@ Header.propTypes = {
   searchComponent: PropTypes.node,
   /** Applies the chosen theme to the entire header. */
   theme: PropTypes.object,
-  /** Specifies alternate versions of the header */
-  variant: PropTypes.oneOf(["search"]),
   /** Allows for use of the `data-testid` attribute for testing. */
   dataTestId: PropTypes.string,
   /** Allows you to pass in child components to user dropdown menu */
